+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - Add support for custom map types (#410)
 - Upgrade to protobuf 3.7.0
 - Removed deprecated com.trueaccord symbols
+- Added `scalapb.GeneratedSealedOneof` as a base trait for all sealed oneofs.
+- Added `sealed_oneof_extends` option to allow adding base classes for sealed oneofs.
 
 ## [0.8.4](https://github.com/scalapb/ScalaPB/tree/v0.8.4)
 [Full Changelog](https://github.com/scalapb/ScalaPB/compare/v0.8.3...v0.8.4)